Update Nexport plugins with additional integration
Case 130255:
- Add SectionCeus property to product mapping
- Show CEUs information if the product is mapped with a Nexport section

Case 130251, Case 132194, Case 132196, Case 132197, Case 132653:
- Add group memberships to mapping
- Add migration to link group memberships and product mapping
- Add group memberships to API call

Case 128632, Case 130257, Case 130324, Case 130334: Initial work on background synchronization task

Case 132368: Add My Trainings page that lists the organizations that Nexport users have access to

Case 132400: Add custom login behavior to support Nexport login credential

Case 132650, Case 132651: Add auto redeeming purchase option to mapping and automatically redeem the purchase during the invoice processing work

+++ src/Plugins/Nop.Plugin.Misc.Nexport/Models/MapProductToNexportProductListModel.cs
@@ -29,6 +29,6 @@ namespace Nop.Plugin.Misc.Nexport.Models
 
         public DateTime? UtcEndDate { get; set; }
 
-        public int? AccessTimeLimit { get; set; }
+        public long? AccessTimeLimit { get; set; }
     }
 }
